customer controller

+++ src/main/java/edu/MosBurgers/repo/CustomerRepository.java
@@ -4,5 +4,5 @@ import edu.MosBurgers.entity.Customer;
 import edu.MosBurgers.entity.FoodItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
 }
